fix(formula): split complex.ts
