+++ packages/engine-formula/src/functions/engineering/imargument/index.ts
@@ -19,7 +19,7 @@ import { ErrorValueObject } from '../../../engine/value-object/base-value-object
 import { BaseFunction } from '../../base-function';
 import { NumberValueObject } from '../../../engine/value-object/primitive-object';
 import { checkVariantsErrorIsArrayOrBoolean } from '../../../engine/utils/check-variant-error';
-import { Complex } from '../../../basics/engineering';
+import { Complex } from '../../../basics/complex';
 import { ErrorType } from '../../../basics/error-type';
 
 export class Imargument extends BaseFunction {
